+++ pysteps/steps/API_2/sim.py
@@ -1858,14 +1858,14 @@ class Simulation(nutils.NamedObject, nutils.StepsWrapperObject, nutils.AdvancedP
             else:
                 raise Exception(f'Expected a ResultSelector object, got {rs} instead.')
 
-    def toDB(self, dbh, uid, **kwargs):
+    def toDB(self, dbh, uid=None, **kwargs):
         """Redirect all the added results selectors to a database
 
         :param dbh: The database to which the result selectors should be saved (see e.g.
             :py:class:`steps.API_2.saving.SQLiteDBHandler`).
         :type dbh: :py:class:`steps.API_2.saving.DatabaseHandler`
         :param uid: A unique identifier under which all subsequent runs should be saved. It should
-            not contain any slashes.
+            not contain any slashes. If it is not provided, a default one will be used.
         :type uid: str
         :param kwargs: Any additional parameters that should be saved to the database along with
             the unique identifier. Values are restricted to the documented types.
@@ -1888,6 +1888,8 @@ class Simulation(nutils.NamedObject, nutils.StepsWrapperObject, nutils.AdvancedP
         """
         if not isinstance(dbh, nsaving.DatabaseHandler):
             raise TypeError(f'Expected a DatabaseHandler, got {dbh} instead.')
+        if uid is None:
+            uid = dbh._getDefaultGroupName()
         if '/' in uid:
             raise ValueError(f'The unique run group identifier cannot contain slashes: {uid}')
         # The list of result selectors can be modified if the mesh is distributed